AG-14087 Add AbstractButtonWidget (prep-work for incoming MenuItemWidget)
