+++ packages/ag-charts-community/src/widget/buttonWidget.ts
@@ -1,29 +1,9 @@
-import { createElement, setAttribute } from 'ag-charts-core';
+import { createElement } from 'ag-charts-core';
 
-import { Widget } from './widget';
-import type { WidgetEventMap as EventMap } from './widgetEvents';
+import { AbstractButtonWidget } from './abstractButtonWidget';
 
-type R = ReturnType<Widget['addListener']>;
-
-export class ButtonWidget extends Widget<HTMLButtonElement> {
+export class ButtonWidget extends AbstractButtonWidget<HTMLButtonElement> {
     constructor() {
         super(createElement('button'));
-        this.setEnabled(true);
-    }
-
-    protected override destructor() {
-        // Nothing to destroy.
-    }
-
-    setEnabled(enabled: boolean) {
-        setAttribute(this.elem, 'aria-disabled', !enabled);
-    }
-
-    override addListener<K extends keyof EventMap>(type: K, listener: (ev: EventMap[K], current: this) => unknown): R;
-    override addListener<K extends keyof EventMap>(type: K, listener: (ev: unknown, current: this) => unknown): R {
-        return super.addListener(type, (ev, current: this) => {
-            if ((type === 'click' || type === 'dblclick') && this.isDisabled()) return;
-            listener(ev, current);
-        });
     }
 }
